issue: 4946495 Merge Ultra API gtests by categories
Merge multiple send_receive and migrate test suites into single respective
test suites. This reduces number of files and copy-paste.
