+++ tests/gtest/xlio_ultra_api/xlio_socket_migrate.cc
@@ -15,12 +15,15 @@
 
 #if defined(EXTRA_API_ENABLED) && (EXTRA_API_ENABLED == 1)
 
+#define NUMBER_OF_MESSAGES    1000
+#define MIGRATE_AFTER_SECONDS 10
+
 static int connected_counter = 0;
 static int terminated_counter = 0;
 static int rx_cb_counter = 0;
 static int comp_cb_counter = 0;
 static const char *data_to_send = "I Love XLIO!";
-static int data_bytes_to_be_sent = strlen(data_to_send) * 1000;
+static int data_bytes_to_be_sent = strlen(data_to_send) * NUMBER_OF_MESSAGES;
 static int data_sent_completed = 0;
 static int data_received = 0;
 static struct ibv_pd *pd = NULL;
@@ -34,7 +37,6 @@ public:
     virtual void SetUp()
     {
         errno = EOK;
-        // Reset static variables between test runs
         connected_counter = 0;
         terminated_counter = 0;
         rx_cb_counter = 0;
@@ -48,7 +50,6 @@ public:
     };
     virtual void TearDown()
     {
-        // Clean up memory registration if it exists (parent process only)
         if (mr_buf) {
             ibv_dereg_mr(mr_buf);
             mr_buf = NULL;
@@ -107,7 +108,7 @@ public:
 /**
  * @test ultra_api_socket_migrate.ti_1
  * @brief
- *    Create TCP socket/connect/send(initiator)/receive(target)
+ *    Create TCP socket/connect/send(initiator)/receive(target) with immediate migrate
  * @details
  */
 TEST_F(ultra_api_socket_migrate, ti_1)
@@ -190,6 +191,124 @@ TEST_F(ultra_api_socket_migrate, ti_1)
                 base_send_single_msg(sock, data_to_send, strlen(data_to_send), strlen(data_to_send),
                                      0, mr_buf, sndbuf);
                 sent_bytes += strlen(data_to_send);
+            }
+        }
+
+        base_wait_for_delayed_acks(group);
+
+        ASSERT_EQ(data_sent_completed, data_bytes_to_be_sent);
+
+        barrier_fork(pid, true); // wait for child to accept + receive last ack
+
+        base_destroy_socket(sock);
+        while (terminated_counter < 1) {
+            xlio_api->xlio_poll_group_poll(group);
+        }
+
+        destroy_poll_group(group);
+
+        wait_fork(pid);
+    }
+}
+
+/**
+ * @test ultra_api_socket_migrate.ti_2
+ * @brief
+ *    Create TCP socket/connect/send(initiator)/receive(target) with delayed migrate
+ * @details
+ */
+TEST_F(ultra_api_socket_migrate, ti_2)
+{
+    int rc;
+    int pid = fork();
+    ultra_api_base::SetUp();
+    xlio_poll_group_t group;
+    xlio_socket_t sock;
+
+    base_create_poll_group(&group, &socket_event_cb, &socket_comp_cb, &socket_rx_cb,
+                           &socket_accept_cb);
+    xlio_socket_attr sattr = {
+        .flags = 0,
+        .domain = server_addr.addr.sa_family,
+        .group = group,
+        .userdata_sq = 0,
+    };
+    if (pid == 0) {
+        xlio_poll_group_t group_2;
+        base_create_poll_group(&group_2, &socket_event_cb, &socket_comp_cb, &socket_rx_cb,
+                               &socket_accept_cb);
+
+        base_create_socket(&sattr, &sock);
+
+        rc = xlio_api->xlio_socket_bind(sock, (struct sockaddr *)&server_addr, sizeof(server_addr));
+        ASSERT_EQ(0, rc);
+
+        rc = xlio_api->xlio_socket_listen(sock);
+        ASSERT_EQ(0, rc);
+
+        barrier_fork(pid, true);
+
+        bool is_detached = false;
+        struct timespec start_time_detach;
+        struct timespec start_time_now;
+        while (data_received < data_bytes_to_be_sent) {
+            xlio_api->xlio_poll_group_poll(group);
+            xlio_api->xlio_poll_group_poll(group_2);
+            // add timer to attach after 10 seconds of detach
+            if (do_migrate) {
+                if (!is_detached) {
+                    clock_gettime(CLOCK_MONOTONIC, &start_time_detach);
+                    xlio_api->xlio_socket_detach_group(accepted_sockets.back());
+                    is_detached = true;
+                }
+                clock_gettime(CLOCK_MONOTONIC, &start_time_now);
+                if (start_time_now.tv_sec - start_time_detach.tv_sec > MIGRATE_AFTER_SECONDS) {
+                    xlio_api->xlio_socket_attach_group(accepted_sockets.back(), group_2);
+                    do_migrate = false;
+                }
+            }
+        }
+
+        base_wait_for_delayed_acks(group);
+
+        ASSERT_EQ(data_received, data_bytes_to_be_sent);
+
+        barrier_fork(pid, true);
+
+        base_destroy_socket(sock);
+        base_cleanup_accepted_sockets(accepted_sockets);
+        while (terminated_counter < 1) {
+            xlio_api->xlio_poll_group_poll(group);
+        }
+
+        if (mr_buf) {
+            ibv_dereg_mr(mr_buf);
+            mr_buf = NULL;
+        }
+
+        destroy_poll_group(group);
+        destroy_poll_group(group_2);
+        exit(testing::Test::HasFailure());
+    } else {
+        base_create_socket(&sattr, &sock);
+
+        rc = xlio_api->xlio_socket_bind(sock, (struct sockaddr *)&client_addr, sizeof(client_addr));
+        ASSERT_EQ(0, rc);
+
+        barrier_fork(pid, true); // wait for child to bind and listen
+
+        rc = xlio_api->xlio_socket_connect(sock, (struct sockaddr *)&server_addr,
+                                           sizeof(server_addr));
+        ASSERT_EQ(0, rc);
+
+        int sent_bytes = 0;
+        while (data_sent_completed < data_bytes_to_be_sent) {
+            xlio_api->xlio_poll_group_poll(group);
+            if (connected_counter > 0 && sent_bytes < data_bytes_to_be_sent) {
+                base_send_single_msg(sock, data_to_send, strlen(data_to_send), strlen(data_to_send),
+                                     0, mr_buf, sndbuf);
+                sent_bytes += strlen(data_to_send);
+                usleep(MIGRATE_AFTER_SECONDS * 1000000 / NUMBER_OF_MESSAGES);
             }
         }
